completely overhauled and simplified routing pkg
