--- routing/consumer/consumer_test.go
@@ -1,58 +1,0 @@
-package consumer
-
-import (
-	"testing"
-	"time"
-
-	"mycelia/boot"
-	"mycelia/commands"
-	"mycelia/test"
-)
-
-var _ = boot.RuntimeCfg // REQUIRED for global config values.
-
-func TestConsumer_ConsumeMessage_SetsResolvedAndSendsBody(t *testing.T) {
-	addr, gotBody, stop := test.MockOneWayServer(t)
-	t.Cleanup(stop)
-
-	c := NewConsumer(addr)
-	msg := &commands.SendMessage{
-		ID:    "abc",
-		Route: "orders.created",
-		Body:  "hello world",
-		// Status starts non-resolved; we only assert it becomes StatusResolved
-	}
-
-	c.ConsumeMessage(msg)
-
-	// Assert status flipped to resolved.
-	if msg.Status != commands.StatusResolved {
-		t.Fatalf("status not set to resolved, got: %v", msg.Status)
-	}
-
-	// Assert the body was actually sent over the wire.
-	select {
-	case b := <-gotBody:
-		if b != "hello world" {
-			t.Fatalf("sent body mismatch: want %q, got %q", "hello world", b)
-		}
-	case <-time.After(200 * time.Millisecond):
-		t.Fatalf("server did not receive body in time")
-	}
-}
-
-func TestConsumer_ConsumeMessage_DialFailure_DoesNotSetResolved(t *testing.T) {
-	// Port 0 is not dialable; this should fail fast and NOT set StatusResolved.
-	c := NewConsumer("127.0.0.1:0")
-	msg := &commands.SendMessage{
-		ID:    "x",
-		Route: "r",
-		Body:  "payload",
-	}
-
-	c.ConsumeMessage(msg)
-
-	if msg.Status == commands.StatusResolved {
-		t.Fatalf("status should not be resolved on dial failure")
-	}
-}
